+++ test/scrollbar.test.js
@@ -1,72 +1,4 @@
-const _ = require('lodash');
-const postcss = require('postcss');
-const snapshotDiff = require('snapshot-diff');
-const tailwindcss = require('tailwindcss');
-const path = require('path');
-
-const scrollbarPlugin = require('..');
-
-/**
- * Generates the CSS for the plugin
- *
- * @see https://www.oliverdavies.uk/blog/testing-tailwind-css-plugins-jest
- * @param {object} config Tailwind config options to pass to tailwind
- * @returns {string} The CSS generated from the plugin using the provided config
- */
-const generatePluginCss = async (config = {}) => {
-  const { warn } = console;
-  console.warn = () => {}; // eslint-disable-line no-console
-
-  const tailwindConfig = _.merge({
-    theme: {
-      colors: {
-        black: '#000000',
-        indigo: {
-          DEFAULT: '#5c6ac4',
-          dark: '#202e78'
-        }
-      }
-    },
-    corePlugins: [],
-    plugins: [scrollbarPlugin]
-  }, config);
-
-  const result = await postcss(tailwindcss(tailwindConfig))
-    .process('@tailwind utilities;', { from: undefined });
-
-  console.warn = warn; // eslint-disable-line no-console
-
-  return result.css;
-};
-
-/**
- * Generates a diff between a default tailwind run and one customized with a
- * config
- *
- * @see https://github.com/tailwindlabs/tailwindcss-typography/blob/master/src/index.test.js
- * @param {object} config The config to diff against
- * @returns {string} The diff between the configured tailwind run and the default
- */
-async function diffOnly(config = {}) {
-  const [before, after] = await Promise.all([
-    generatePluginCss(),
-    generatePluginCss(config)
-  ]);
-
-  return `\n\n${snapshotDiff(before, after, {
-    aAnnotation: '__REMOVE_ME__',
-    bAnnotation: '__REMOVE_ME__',
-    contextLines: 0
-  })
-    .replace(/\n\n@@([^@@]*)@@/g, '') // Top level @@ signs
-    .replace(/@@([^@@]*)@@/g, '\n---\n') // In between @@ signs
-    .replace(/[-+] __REMOVE_ME__\n/g, '')
-    .replace(/Snapshot Diff:\n/g, '')
-    .replace(/"/g, '\'')
-    .split('\n')
-    .map(line => `  ${line}`.trimEnd())
-    .join('\n')}\n\n`;
-}
+const { generatePluginCss, diffOnly } = require('./util');
 
 test('it generates scrollbar utilities', async () => {
   const css = await generatePluginCss();
@@ -178,55 +110,6 @@ test('it generates scrollbar utilities', async () => {
 
     .hover\\\\:scrollbar-thumb-indigo-dark::-webkit-scrollbar-thumb:hover {
       --scrollbar-thumb: #202e78;
-    }"
-`);
-});
-
-test('it works in jit mode', async () => {
-  const css = await generatePluginCss({
-    mode: 'jit',
-    purge: [path.resolve(__dirname, './jit-mode.html')],
-    corePlugins: ['textColor']
-  });
-
-  expect(css).toMatchInlineSnapshot(`
-    ".scrollbar {
-        --scrollbar-track: initial;
-        --scrollbar-thumb: initial;
-        scrollbar-color: var(--scrollbar-thumb) var(--scrollbar-track);
-        overflow: overlay;
-    }
-    .scrollbar.overflow-x-hidden {
-        overflow-x: hidden;
-    }
-    .scrollbar.overflow-y-hidden {
-        overflow-y: hidden;
-    }
-    .scrollbar::-webkit-scrollbar-track {
-        background-color: var(--scrollbar-track);
-    }
-    .scrollbar::-webkit-scrollbar-thumb {
-        background-color: var(--scrollbar-thumb);
-    }
-    .scrollbar {
-        scrollbar-width: auto;
-    }
-    .scrollbar::-webkit-scrollbar {
-        width: 16px;
-        height: 16px;
-    }
-    .scrollbar-thumb-indigo {
-        --scrollbar-thumb: #5c6ac4;
-    }
-    .hover\\\\:text-black:hover {
-        --tw-text-opacity: 1;
-        color: rgba(0, 0, 0, var(--tw-text-opacity));
-    }
-    .hover\\\\:scrollbar-track-black::-webkit-scrollbar-track:hover {
-        --scrollbar-track: #000000;
-    }
-    .hover\\\\:scrollbar-thumb-indigo-dark::-webkit-scrollbar-thumb:hover {
-        --scrollbar-thumb: #202e78;
     }"
 `);
 });
